libs/common/pyproject.toml
[project]
name = "llamatrade-common"
version = "0.1.0"
description = "Shared utilities and models for LlamaTrade services"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pydantic>=2.5.0",
    "pydantic[email]>=2.5.0",
    "PyJWT>=2.8.0",
    "cryptography>=41.0.0",
    "prometheus-client>=0.19.0",
    "fastapi>=0.109.0",
    "httpx>=0.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
]
integration = [
    "testcontainers[postgres,redis]>=4.0.0",
    "pytest-xdist>=3.5.0",
    "pytest-timeout>=2.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["llamatrade_common"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=llamatrade_common --cov-report=term-missing --cov-fail-under=80"

[tool.coverage.run]
source = ["llamatrade_common"]
branch = true

[tool.coverage.report]
fail_under = 80
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[tool.ruff]
line-length = 100
target-version = "py314"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pyright]
venvPath = "../.."
venv = ".venv"
pythonVersion = "3.14"
